pyproject.toml
[tool.poetry]
name = "apple-music-mcp"
version = "0.1.0"
description = "Hybrid MCP for Apple Music leveraging MusicKit and AppleScript"
authors = ["InWonder <dev@inwonder.example>"]
readme = "README.md"
packages = [{ include = "handlers" }, { include = "utils" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.5.0"
PyJWT = "^2.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
